Убрал пробелы между оператором цикла и скобкой. Добавил пробел в конце программы.
